requirements.txt
fastApi
uvicorn
psycopg2-binary
pylint
pydantic
python-dotenv